Move Router to a separate component and delete unnecessary imports
